+++ modules/time.c
@@ -1,4 +1,4 @@
-#include "time/time.h"
+#include "modules/time.h"
 $int time$$time () {
     struct timespec ts;
     if (clock_gettime(CLOCK_REALTIME, &ts) == -1) {
